+++ database/migrations/2020_01_01_000008_create_assigned_roles_table.php
@@ -17,10 +17,10 @@ class CreateAssignedRolesTable extends Migration
     {
         Schema::create(config('cortex.auth.tables.assigned_roles'), function (Blueprint $table) {
             // Columns
-            $table->bigInteger('role_id')->unsigned();
-            $table->bigInteger('entity_id')->unsigned();
+            $table->integer('role_id')->unsigned();
+            $table->integer('entity_id')->unsigned();
             $table->string('entity_type', 150);
-            $table->bigInteger('scope')->nullable();
+            $table->integer('scope')->nullable();
 
             // Indexes
             $table->index(['scope']);
